requirements.txt
xgboost
scikit-learn
pandas
numpy
wandb
joblib
google-cloud-storage
flask
fastapi
uvicorn
requests